Replace NTP with Chrony
NTP does not handle the Sleep and Hibernate states on the *host* gracefully;
the clock on the guest becomes stale and is not updated regularly (if ever).

Replacing NTP with Chrony fixes the issue, provided chronyd is configured
to allow unlimited catching-up via "makestep 1 -1" (the second value
represents unlimited catch-up, which is necessary when the host has been
sleeping or hibernating for any period of time longer than a few seconds).

Resolves https://github.com/laravel/homestead/issues/1300