set the MySQL timezone to explicitly use UTC
MySQL by default uses the "SYSTEM" timezone. on most servers this is UTC, so everything is fine, but on Homestead it may sync with the host OS to a non-UTC timezone. We could try and change the guest OS timezone, but this might be difficult across the different providers and user preferences. Instead, we'll **explicitly** set the MySQL timezone to UTC.

`timestamp` column types will automatically try to convert any values they receive to UTC. most applications are already handling dates in the code, and delivering UTC values to MySQL. this can cause issues if MySQL receives a value that does not exist in it's given timezone.

for example, we recently had daylight savings time taking us from CST to CDT. The datetime between 2024-03-10 02:00:00 - 2024-03-10 02:59:59 technically does not exist, so if you try to feed this value to MySQL, it will reject the data.